Refactor networking namespace to Protocols
Replaced Nalix.Common.Networking.Packets with Nalix.Common.Networking.Protocols across multiple files, including updating using directives and the namespace in ProtocolOpCode.cs, to better reflect protocol-related functionality and improve code organization.
